flattened a little bit, still feels pretty complex
